Added simple tutorial for first time users
Added a tooltip indicating that users can start formulating by entering an ingredient name in the add ingredient input. Selecting the input will remove the tooltip and be marked in Cookies so the tutorial will not show up on every page load.

Currently this is toggleable by clicking the magnifying glass for testing purposes.

Functionality should be extendable to a proper tutorial in the future by saving a step counter in the cookies instead of just a boolean value with a -1 value hiding all tutorial steps.