Disabled auto update ingredient cost in Formulate
Found issue where cost from ingredients in db would always override costs added in the Formulating view. This would result in costs set to 0 overriding actual costs which is not intended. I added an extra check on 0 values as that is most commonly not intended as ovverride if there is and actual cost entered in the Formula. I then decided that it should probably be prompted to the user, so I decided to deactivate the functionality entirely until a prompt is implemented.