Added Ingredient info
Added info field to the Ingredient type so users can store various info on their ingredients.

Added actual functionality to the IngredientBuilder and replaced instances of constructor being called directly. This to consolidate the code and make sure that changes, like adding the info field, only need to be added in one place.